WFLY-9363 Fix failing clustering EJB2 tests
* make tests run by default
* bunch of simplifications
