+++ testsuite/integration/clustering/pom.xml
@@ -267,31 +267,6 @@
                                 </configuration>
                             </execution>
 
-                            <!-- Multinode clustering tests with custom containers with unmanaged deployments -->
-                            <!-- Tests which are placed in clustering module but they're not run by default -->
-                            <execution>
-                                <id>ts.surefire.clustering.multinode-manual-tcp-sync-extended</id>
-                                <phase>test</phase> 
-                                <goals><goal>test</goal></goals>
-                                <configuration>
-                                    <!-- Tests to execute. -->
-                                    <includes>
-                                        <include>org/jboss/as/test/clustering/extended/**/*TestCase.java</include>
-                                    </includes>
-
-                                    <!-- Parameters to test cases. -->
-                                    <systemPropertyVariables combine.children="append">
-                                        <arquillian.launch>clustering-all</arquillian.launch>
-                                        <jboss.server.config.file.name>standalone-ha.xml</jboss.server.config.file.name>
-                                        <cacheMode>SYNC</cacheMode>
-                                        <stack>tcp</stack>
-                                    </systemPropertyVariables>
-
-                                    <!-- Differentiate surefire TCP and UDP reports. -->
-                                    <reportNameSuffix>SYNC-tcp-extended</reportNameSuffix>
-                                </configuration>
-                            </execution>
-
                             <!-- TUNNEL-ASYNC -->
                             <execution>
                                 <id>ts.surefire.clustering.multinode-tunnel-async</id>
