+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/ejb2/stateless/StatelessBean.java
@@ -1,0 +1,42 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2011, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+package org.jboss.as.test.clustering.cluster.ejb2.stateless;
+
+import javax.ejb.Remote;
+import javax.ejb.RemoteHome;
+import javax.ejb.SessionBean;
+import javax.ejb.Stateless;
+
+import org.jboss.as.test.clustering.cluster.ejb2.stateless.bean.StatelessBeanBase;
+import org.jboss.as.test.clustering.cluster.ejb2.stateless.bean.StatelessRemote;
+import org.jboss.as.test.clustering.cluster.ejb2.stateless.bean.StatelessRemoteHome;
+
+/**
+ * @author Ondrej Chaloupka
+ */
+@Stateless
+@RemoteHome(StatelessRemoteHome.class)
+@Remote(StatelessRemote.class)
+public class StatelessBean extends StatelessBeanBase implements SessionBean {
+    private static final long serialVersionUID = 1L;
+}
