--- testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/extended/ejb2/stateful/remote/failover/dd/ejb-jar.xml
@@ -1,36 +1,0 @@
-<?xml version='1.0' ?>
-<!DOCTYPE ejb-jar PUBLIC "-//Sun Microsystems, Inc.//DTD Enterprise JavaBeans 2.0//EN"
-                         "http://java.sun.com/dtd/ejb-jar_2_0.dtd">
-<!--
-  ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2012, Red Hat, Inc., and individual contributors
-  ~ as indicated by the @author tags. See the copyright.txt file in the
-  ~ distribution for a full listing of individual contributors.
-  ~
-  ~ This is free software; you can redistribute it and/or modify it
-  ~ under the terms of the GNU Lesser General Public License as
-  ~ published by the Free Software Foundation; either version 2.1 of
-  ~ the License, or (at your option) any later version.
-  ~
-  ~ This software is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  ~ Lesser General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Lesser General Public
-  ~ License along with this software; if not, write to the Free
-  ~ Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-  ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-  -->
-<ejb-jar>
-    <enterprise-beans>
-        <session>
-            <ejb-name>CounterBean</ejb-name>
-            <home>org.jboss.as.test.clustering.extended.ejb2.stateful.remote.failover.CounterRemoteHome</home>
-            <remote>org.jboss.as.test.clustering.extended.ejb2.stateful.remote.failover.CounterRemote</remote>
-            <ejb-class>org.jboss.as.test.clustering.extended.ejb2.stateful.remote.failover.dd.CounterBeanDD</ejb-class>
-            <session-type>Stateful</session-type>
-            <transaction-type>Container</transaction-type>
-        </session>
-    </enterprise-beans>
-</ejb-jar>
